refactor: move truncate() from table.rs to health.rs
table.rs no longer uses truncation — move the function to its only
remaining consumer to keep the table module clean.
